Simplify appBar to reduce shared logic, extract breadcrumbs to helper
