+++ packages/web-app-files/src/components/TrashBin.vue
@@ -2,6 +2,11 @@
   <div>
     <app-loading-spinner v-if="loadResourcesTask.isRunning" />
     <template v-else>
+      <app-bar
+        :breadcrumbs="breadcrumbs"
+        :breadcrumbs-context-actions-displayed="breadcrumbsContextActionsDisplayed"
+        :has-bulk-actions="true"
+      />
       <no-content-message
         v-if="isEmpty"
         id="files-trashbin-empty"
@@ -48,6 +53,7 @@
 
 <script>
 import { mapGetters, mapMutations, mapState } from 'vuex'
+import AppBar from './AppBar/AppBar.vue'
 import ResourceTable from './FilesList/ResourceTable.vue'
 
 import MixinFilesListFilter from '../mixins/filesListFilter'
@@ -66,6 +72,7 @@ export default {
   name: 'TrashBin',
 
   components: {
+    AppBar,
     ResourceTable,
     AppLoadingSpinner,
     NoContentMessage,
@@ -76,6 +83,8 @@ export default {
 
   mixins: [MixinResources, MixinMountSideBar, MixinFilesListFilter],
   props: {
+    breadcrumbs: { type: Array, default: () => [] },
+    breadcrumbsContextActionsDisplayed: { type: Boolean, default: false },
     noContentMessage: {
       type: String,
       required: false,
